refactor: flatten four single-class utils packages into modules
lineage, id_generation, field_management, and correlation were each a
directory with a five-line __init__ re-exporting one class from one
file. Collapse each into a plain module of the same name — package-
level imports (from agent_actions.utils.lineage import LineageBuilder)
are unchanged; only inner-module paths moved. transformation/ keeps
its package shape: it has a real strategies/ subpackage.
