+++ .changes/unreleased/Under the Hood-20260807-220000.yaml
@@ -1,0 +1,3 @@
+kind: Under the Hood
+body: "Flatten four single-class utils packages (lineage, id_generation, field_management, correlation) into plain modules — package-level imports are unchanged; only inner-module paths moved. No behavior change."
+time: 2026-08-07T22:00:00Z
