+++ agent_actions/record/ARCHITECTURE.md
@@ -234,7 +234,7 @@ History is capped at `STATE_HISTORY_CAP` (64 entries). When overflow occurs, the
 
 ## Source GUID Generation
 
-Source GUIDs are **not** generated by the record module — they are assigned upstream by `IDGenerator` (`utils/id_generation/generator.py`) and carried through by `_carry_persistent_fields()`.
+Source GUIDs are **not** generated by the record module — they are assigned upstream by `IDGenerator` (`utils/id_generation.py`) and carried through by `_carry_persistent_fields()`.
 
 ```
 First-stage records:
