--- agent_actions/utils/field_management/__init__.py
@@ -1,5 +1,0 @@
-"""Field management utilities for processors."""
-
-from .manager import FieldManager
-
-__all__ = ["FieldManager"]
