--- agent_actions/utils/id_generation/_MANIFEST.md
@@ -1,13 +1,0 @@
-# ID Generation Manifest
-
-## Overview
-
-Centralized UUID helpers for `target_id`, `node_id`, and deterministic `source_guid`
-generation used by processors, UDF registries, and field managers.
-
-## Modules
-
-| Name | Type | Description | Signals |
-|------|------|-------------|---------|
-| `generator.py` | Module | `IDGenerator` methods that emit UUID4 targets, `{action}_{uuid}` node IDs, and deterministic UUID5 source GUIDs. | `field_management`, `lineage` |
-| `IDGenerator` | Class | Static helpers `generate_target_id`, `generate_node_id`, `derive_source_guid` (deterministic first-stage identity), `generate_source_guid` (random, for new entities), and `generate_content_hash`. | `field_management`, `lineage` |
