--- agent_actions/utils/lineage/__init__.py
@@ -1,5 +1,0 @@
-"""Lineage tracking utilities for processors."""
-
-from .builder import LineageBuilder
-
-__all__ = ["LineageBuilder"]
